Added ConfigLong

+++ src/main/java/com/feed_the_beast/ftblib/lib/config/ConfigTimer.java
@@ -79,7 +79,7 @@ public class ConfigTimer extends ConfigValue implements LongSupplier
 	}
 
 	@Override
-	public double getDouble()
+	public long getLong()
 	{
 		return getTimer().ticks();
 	}
